# packages/types/openapi.yaml
openapi: 3.1.0
info:
  title: VisaOnTrack v2 API
  version: 0.2.5
  description: |
    Two-sided marketplace API connecting visa seekers with vetted service providers in Thailand.

    ## Authentication
    This API uses JWT tokens stored in HttpOnly cookies. Include credentials with requests.

    ## Entitlements & Quotas
    Provider plans (Free/Pro/Agency) enforce entitlements at the API middleware level.
    Over-quota actions return structured errors (403 EntitlementExceeded, 429 Throttled).

    ## Versioning
    Contract version: 0.2.5 (semver) - RFC-005: Consultation offerings
    Breaking changes will increment major version and require client regeneration.
  
  contact:
    name: VisaOnTrack API Support
  
  license:
    name: Proprietary
servers:
  - url: https://api.visaontrack.com
    description: Production
  - url: https://api-staging.visaontrack.com
    description: Staging
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /requests/{id}/messages:
    get:
      tags:
        - messages
      summary: List messages
      description: Get messages for a request (participants only)
      operationId: listMessages
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Message list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    
    post:
      tags:
        - messages
      summary: Send message
      description: Send message in request thread (participants only)
      operationId: sendMessage
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/AttachmentLimitExceeded'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
...
  /requests/{id}/quotes:
    get:
      tags:
        - quotes
      summary: List quotes for request
      description: Get all quotes/proposals for a request (seeker sees all, provider sees own)
      operationId: listQuotesForRequest
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Quote list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Quote'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - quotes
      summary: Submit quote
      description: 'Submit quote for request (PROVIDER role, entitlement check: quotes.monthly.max)'
      operationId: submitQuote
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuoteRequest'
      responses:
        '201':
          description: Quote submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/EntitlementExceeded'
        '429':
          $ref: '#/components/responses/Throttled'
  /quotes/{id}:
    get:
      tags:
        - quotes
      summary: Get quote by ID
      description: Get quote details
      operationId: getQuote
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Quote details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '404':
          $ref: '#/components/responses/NotFound'
    
    patch:
      tags:
        - quotes
      summary: Update quote
      description: 'Update quote status (provider owner or seeker can accept/decline)'
      operationId: updateQuote
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuoteRequest'
      responses:
        '200':
          description: Quote updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== CHECKOUT ====================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token in HttpOnly cookie
  parameters:
    Page:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
...
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 'UNAUTHORIZED'
            message: 'Authentication required'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 'FORBIDDEN'
            message: 'Insufficient permissions'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 'NOT_FOUND'
            message: 'Resource not found'
    EntitlementExceeded:
      description: Entitlement limit exceeded (e.g., quotes.monthly.max, packages.max)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EntitlementExceededError'
          example:
            code: 'ENTITLEMENT_EXCEEDED'
            message: 'Monthly quote limit exceeded'
            entitlement: 'quotes.monthly.max'
            limit: 50
            used: 50
    Throttled:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ThrottledError'
          example:
            code: 'THROTTLED'
            message: 'Rate limit exceeded'
            retryAfter: 3600
    PayloadTooLarge:
      description: Request payload too large (attachment size limit)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 'PAYLOAD_TOO_LARGE'
            message: 'Attachment size exceeds limit'
    AttachmentLimitExceeded:
      description: Attachment count limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 'ATTACHMENT_LIMIT_EXCEEDED'
            message: 'Attachment count limit exceeded'
...
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
    EntitlementExceededError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            entitlement:
              type: string
              description: Entitlement key (e.g., 'quotes.monthly.max')
            limit:
              type: number
              description: Entitlement limit
            used:
              type: number
              description: Current usage
    ThrottledError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            retryAfter:
              type: integer
              description: Seconds until retry allowed
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    
    # ==================== AUTH ====================
...
    Message:
      type: object
      description: Message schema per Section 3 data model
      required:
        - id
        - requestId
        - senderId
        - body
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        requestId:
          type: string
          format: uuid
          description: Foreign key to Request
        senderId:
          type: string
          format: uuid
          description: Foreign key to User (sender)
        body:
          type: string
          minLength: 1
          maxLength: 10000
          description: Message content
        createdAt:
          type: string
          format: date-time
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
          description: Attachments linked to this message
          nullable: true
    MessageListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CreateMessageRequest:
      type: object
      description: Send message in request thread (participants only, attachment limits enforced)
      required:
        - body
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 10000
        attachmentIds:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 10
          description: 'Attachment IDs from POST /messages/attachments (entitlement check: attachments.maxSizeMB)'
...
    Quote:
      type: object
      description: Quote schema per Section 3 data model
      required:
        - id
        - requestId
        - providerId
        - items
        - totalTHB
        - etaDays
        - status
      properties:
        id:
          type: string
          format: uuid
        requestId:
          type: string
          format: uuid
          description: Foreign key to Request
        providerId:
          type: string
          format: uuid
          description: Foreign key to ProviderProfile
        items:
          type: array
          items:
            $ref: '#/components/schemas/QuoteItem'
          minItems: 1
          description: Quote line items
        totalTHB:
          type: number
          minimum: 0
          description: Total quote amount in Thai Baht
        etaDays:
          type: integer
          minimum: 1
          description: Estimated completion time in days
        terms:
          type: string
          maxLength: 2000
          nullable: true
          description: Terms and conditions
        status:
          $ref: '#/components/schemas/QuoteStatus'
        validUntil:
          type: string
          format: date-time
          nullable: true
          description: Quote expiration timestamp
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        provider:
          type: object
          description: Provider information (included in list/get responses)
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            businessName:
              type: string
            businessType:
              type: string
              nullable: true
            yearsExperience:
              type: integer
              nullable: true
            verifiedAt:
              type: string
              format: date-time
              nullable: true
    QuoteItem:
      type: object
      description: Quote line item
      required:
        - title
        - quantity
        - priceTHB
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Item description
        quantity:
          type: integer
          minimum: 1
          default: 1
        priceTHB:
          type: number
          minimum: 0
          description: Unit price in Thai Baht
        subtotalTHB:
          type: number
          minimum: 0
          description: Calculated subtotal (quantity * priceTHB)
          readOnly: true
    QuoteStatus:
      type: string
      enum:
        - PENDING
        - ACCEPTED
        - DECLINED
        - EXPIRED
    CreateQuoteRequest:
      type: object
      description: 'Submit quote for request (PROVIDER role, entitlement check: quotes.monthly.max)'
      required:
        - items
        - totalTHB
        - etaDays
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/QuoteItem'
          minItems: 1
        totalTHB:
          type: number
          minimum: 0
          description: Total amount in Thai Baht (should match sum of items)
        etaDays:
          type: integer
          minimum: 1
        terms:
          type: string
          maxLength: 2000
          nullable: true
    UpdateQuoteRequest:
      type: object
      description: 'Update quote status (provider owner or seeker can accept/decline)'
      properties:
        status:
          $ref: '#/components/schemas/QuoteStatus'
          description: Status change (seeker can set ACCEPTED/DECLINED, provider can update)
        terms:
          type: string
          maxLength: 2000
          description: Update terms (provider only)
    
    # ==================== CHECKOUT ====================
...
    Attachment:
      type: object
      description: Attachment schema per Section 3 data model
      required:
        - id
        - ownerUserId
        - key
        - mime
        - size
      properties:
        id:
          type: string
          format: uuid
        ownerUserId:
          type: string
          format: uuid
          description: Foreign key to User (owner)
        requestId:
          type: string
          format: uuid
          nullable: true
          description: Foreign key to Request (if attached to request)
        orderId:
          type: string
          format: uuid
          nullable: true
          description: Foreign key to Order (if attached to order)
        key:
          type: string
          description: Storage key (S3/R2 object key)
          example: 'attachments/xxx-xxx-xxx/file.pdf'
        mime:
          type: string
          description: MIME type
          example: 'application/pdf'
        size:
          type: integer
          minimum: 0
          description: File size in bytes
    
    # ==================== BILLING ====================
...